Update: south facing stabs, indicators silkscreen
- Due to a recommendation by anthony001, the spacebar and numpad 0
stabilizers were made south-facing due to that being the default
configuration.

- Also, the indicator LEDs were marked SCROLL/NUM/CAPS LOCK in the
silkscreen. This does not make sense as the order of the LEDs is
customizable and programmable to the liking of the user. Now the
silkscreen marks "I1/2/3" for "Indicator 1/2/3" and also mark the MCU
pin attach to each indicator light.