Update: corrected pads on SMD RGB LED footprint
- The LED_RGB_SMD_6028 footprint was really crude in the sense that it
was developed with pads that overlapped the center hole. This caused
KiCad's DRC to fire up four errors for each unit used in the PCB; in the
TheKeenLadies PCB, this amounted to almos 2000 errors. The pads were now
fine-tunes to the exact shape of the center hole, with a 6 mil spacing
between them.